+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Changed in v13
 
+* Updated for 8.0.1
 * Left-click on rares to open up the group finder searching for them by name (thanks znf)
 * Many more treasures in Argus (thanks znf)
 * Missing from Azsuna